+++ tests/integrators/algorithms/instrumented/generic_erk.py
@@ -4,7 +4,7 @@ from typing import Callable, Optional
 
 import attrs
 from attrs import validators
-from numba import cuda, int16, int32
+from numba import cuda, int32, int32
 
 from cubie._utils import PrecisionDType
 from cubie.cuda_simsafe import activemask, all_sync
@@ -213,8 +213,8 @@ class ERKStep(ODEExplicitStep):
                 numba_precision[:, :, ::1],
                 numba_precision,
                 numba_precision,
-                int16,
-                int16,
+                int32,
+                int32,
                 numba_precision[::1],
                 numba_precision[::1],
                 int32[::1],
@@ -346,7 +346,7 @@ class ERKStep(ODEExplicitStep):
             if first_same_as_last and multistage:
                 if not first_step_flag:
                     mask = activemask()
-                    all_threads_accepted = all_sync(mask, accepted_flag != int16(0))
+                    all_threads_accepted = all_sync(mask, accepted_flag != int32(0))
                     use_cached_rhs = all_threads_accepted
             else:
                 use_cached_rhs = False
