+++ tests/integrators/algorithms/test_step_algorithms.py
@@ -5,7 +5,7 @@ from typing import Any, Optional
 
 import numpy as np
 import pytest
-from numba import cuda, from_dtype, int16
+from numba import cuda, from_dtype, int32
 from numpy.testing import assert_allclose
 
 from cubie.integrators.algorithms import get_algorithm_step
@@ -738,8 +738,8 @@ def device_step_results(
         observables_function(state, params_vec, drivers_vec, observables_vec,
                              precision(0.0))
         shared[:] = precision(0.0)
-        first_step_flag = int16(1)
-        accepted_flag = int16(1)
+        first_step_flag = int32(1)
+        accepted_flag = int32(1)
         result = step_function(
             state_vec,
             proposed_vec,
@@ -915,8 +915,8 @@ def _execute_step_twice(
             error_vec_first,
             dt_scalar,
             zero,
-            int16(1),
-            int16(1),
+            int32(1),
+            int32(1),
             shared,
             persistent,
             counters_vec_first,
@@ -944,8 +944,8 @@ def _execute_step_twice(
             error_vec_second,
             dt_scalar,
             dt_scalar,
-            int16(0),
-            int16(1),
+            int32(0),
+            int32(1),
             shared,
             persistent,
             counters_vec_second,
